Move MethodCallInterface and PropertyAssignmentInterface in a sub-namespace
